@@ -2,48 +2,48 @@
     "type": "project",
     "license": "MIT",
     "require": {
-        "php": ">=8.0.2",
+        "php": ">=8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "api-platform/core": "2.6.*",
-        "composer/package-versions-deprecated": "^1.11",
-        "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2.1",
-        "doctrine/doctrine-migrations-bundle": "^3.0",
-        "doctrine/orm": "^2.7",
-        "guzzlehttp/guzzle": "^7.0",
+        "api-platform/core": "^3.0",
+        "composer/package-versions-deprecated": "1.11.99.5",
+        "doctrine/annotations": "^1.13",
+        "doctrine/doctrine-bundle": "^2.7",
+        "doctrine/doctrine-migrations-bundle": "^3.2",
+        "doctrine/orm": "^2.12",
+        "guzzlehttp/guzzle": "^7.4",
         "lexik/jwt-authentication-bundle": "^2.10",
-        "nelmio/cors-bundle": "^2.1",
-        "phpdocumentor/reflection-docblock": "^5.2",
-        "symfony/asset": "6.0.*",
-        "symfony/console": "6.0.*",
-        "symfony/dotenv": "6.0.*",
-        "symfony/expression-language": "6.0.*",
-        "symfony/flex": "^1.3.1",
-        "symfony/framework-bundle": "6.0.*",
-        "symfony/mercure-bundle": "^0.3.2",
-        "symfony/monolog-bundle": "^3.5",
-        "symfony/property-access": "6.0.*",
-        "symfony/property-info": "6.0.*",
-        "symfony/runtime": "6.0.*",
-        "symfony/security-bundle": "6.0.*",
-        "symfony/serializer": "6.0.*",
-        "symfony/translation": "6.0.*",
-        "symfony/twig-bundle": "6.0.*",
-        "symfony/validator": "6.0.*",
-        "symfony/yaml": "6.0.*"
+        "nelmio/cors-bundle": "^2.2",
+        "phpdocumentor/reflection-docblock": "^5.3",
+        "symfony/asset": "6.1.*",
+        "symfony/console": "6.1.*",
+        "symfony/dotenv": "6.1.*",
+        "symfony/expression-language": "6.1.*",
+        "symfony/flex": "^2.2",
+        "symfony/framework-bundle": "6.1.*",
+        "symfony/mercure-bundle": "^0.3.5",
+        "symfony/monolog-bundle": "^3.8",
+        "symfony/property-access": "6.1.*",
+        "symfony/property-info": "6.1.*",
+        "symfony/runtime": "6.1.*",
+        "symfony/security-bundle": "6.1.*",
+        "symfony/serializer": "6.1.*",
+        "symfony/translation": "6.1.*",
+        "symfony/twig-bundle": "6.1.*",
+        "symfony/validator": "6.1.*",
+        "symfony/yaml": "6.1.*"
     },
     "require-dev": {
-        "api-platform/schema-generator": "^4.0",
+        "api-platform/schema-generator": "^5.0",
         "doctrine/doctrine-fixtures-bundle": "^3.4",
-        "symfony/browser-kit": "6.0.*",
-        "symfony/css-selector": "6.0.*",
-        "symfony/debug-bundle": "6.0.*",
-        "symfony/maker-bundle": "^1.21",
-        "symfony/phpunit-bridge": "6.0.*",
-        "symfony/stopwatch": "6.0.*",
-        "symfony/var-dumper": "6.0.*",
-        "symfony/web-profiler-bundle": "6.0.*"
+        "symfony/browser-kit": "6.1.*",
+        "symfony/css-selector": "6.1.*",
+        "symfony/debug-bundle": "6.1.*",
+        "symfony/maker-bundle": "^1.44",
+        "symfony/phpunit-bridge": "6.1.*",
+        "symfony/stopwatch": "6.1.*",
+        "symfony/var-dumper": "6.1.*",
+        "symfony/web-profiler-bundle": "6.1.*"
     },
     "config": {
         "optimize-autoloader": true,
@@ -95,7 +95,8 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "6.0.*"
+            "require": "6.1.*",
+            "docker": false
         }
     }
 }
