@@ -2,44 +2,42 @@
 
 namespace App\Entity;
 
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\Post;
+use ApiPlatform\Metadata\GetCollection;
+use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\Patch;
+use ApiPlatform\Metadata\Put;
+use ApiPlatform\Metadata\Get;
+use ApiPlatform\Metadata\ApiResource;
+use ApiPlatform\Metadata\ApiProperty;
+use ApiPlatform\Doctrine\Orm\Filter\OrderFilter;
+use ApiPlatform\Metadata\ApiFilter;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Validator\Constraints as Assert;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
-use ApiPlatform\Core\Annotation\ApiProperty;
 use Symfony\Component\Serializer\Annotation\Groups;
-use ApiPlatform\Core\Annotation\ApiFilter;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\OrderFilter;
 use App\Filter\SimpleSearchFilter;
-
 /**
  * Class defining entities with data about an Employees
  *
- * @ApiResource(
- *     attributes={"order"={"lastName", "firstName"}},
- *     itemOperations={
- *          "get"={
- *              "normalization_context"={"groups"={"employee_get"}},
- *               "security"="is_granted('ROLE_ADMIN') or object.getUser() == user"
- *          },
- *          "put"={"security_post_denormalize"="is_granted('ROLE_ADMIN') or
-               (object.getUser() == user and previous_object.getUser() == user)"},
- *          "patch"={"security_post_denormalize"="is_granted('ROLE_ADMIN') or
-               (object.getUser() == user and previous_object.getUser() == user)"},
- *          "delete"={"security"="is_granted('ROLE_ADMIN')" }
- *     },
- *     collectionOperations={
- *         "get"={
- *              "normalization_context"={"groups"={"employee_list"}}
- *          },
- *          "post"={"security"="is_granted('ROLE_ADMIN')" }
- *     }
- * )
- * @ApiFilter(OrderFilter::class)
- * @ApiFilter(SimpleSearchFilter::class, properties={"lastName", "firstName", "job", "address", "zipcode", "city"}, arguments={"searchParameterName"="search"})
- * @ORM\Entity
  */
+#[ORM\Entity]
+#[ApiResource(operations: [
+    new Get(normalizationContext: ['groups' => ['employee_get']],
+        security: 'is_granted(\'ROLE_ADMIN\') or object.getUser() == user'),
+    new Put(securityPostDenormalize: 'is_granted(\'ROLE_ADMIN\') or
+               (object.getUser() == user and previous_object.getUser() == user)'),
+    new Patch(securityPostDenormalize: 'is_granted(\'ROLE_ADMIN\') or
+               (object.getUser() == user and previous_object.getUser() == user)'),
+    new Delete(security: 'is_granted(\'ROLE_ADMIN\')'),
+    new GetCollection(normalizationContext: ['groups' => ['employee_list']]),
+    new Post(security: 'is_granted(\'ROLE_ADMIN\')')], order: ['lastName', 'firstName'])
+]
+#[ApiFilter(filterClass: OrderFilter::class)]
+#[ApiFilter(filterClass: SimpleSearchFilter::class,
+    properties: ['lastName', 'firstName', 'job', 'address', 'zipcode', 'city'],
+    arguments: ['searchParameterName' => 'search'])]
 class Employee
 {
     public function __construct()
@@ -49,100 +47,95 @@ class Employee
 
     /**
      * @var int The entity Id
-     *
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
      */
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type:"integer")]
     private $id;
 
     /**
      * @var string
-     * @ORM\Column(nullable=true)
-     * @Assert\Length(max=20)
-     * @Groups({"employee_get"})
      */
+    #[ORM\Column(nullable:true)]
+    #[Assert\Length(max:20)]
+    #[Groups(["employee_get"])]
     private $firstName;
 
     /**
      * @var string
-     * @ORM\Column
-     * @Assert\NotBlank
-     * @Assert\Length(max=80)
-     * @Groups({"employee_get"})
      */
+    #[ORM\Column]
+    #[Assert\NotBlank]
+    #[Assert\Length(max:80)]
+    #[Groups(["employee_get"])]
     private $lastName;
 
     /**
      * @var string
-     * @ORM\Column
-     * @Assert\NotBlank
-     * @Assert\Length(max=40)
-     * @Groups({"employee_get", "employee_list"})
      */
+    #[ORM\Column]
+    #[Assert\NotBlank]
+    #[Assert\Length(max:40)]
+    #[Groups(["employee_get", "employee_list"])]
     private $job;
 
     /**
      * @var string
-     * @ORM\Column
-     * @Assert\NotBlank
-     * @Assert\Length(max=80)
-     * @Groups({"employee_get"})
      */
+    #[ORM\Column]
+    #[Assert\NotBlank]
+    #[Assert\Length(max:80)]
+    #[Groups(["employee_get"])]
     private $address;
 
     /**
      * @var string|null
-     * @ORM\Column(nullable=true)
-     * @Assert\Length(max=10)
-     * @Groups({"employee_get"})
      */
+    #[ORM\Column(nullable:true)]
+    #[Assert\Length(max:10)]
+    #[Groups(["employee_get"])]
     private $zipcode;
 
     /**
      * @var string
-     * @ORM\Column
-     * @Assert\NotBlank
-     * @Assert\Length(max=40)
-     * @Groups({"employee_get"})
      */
+    #[ORM\Column]
+    #[Assert\NotBlank]
+    #[Assert\Length(max:40)]
+    #[Groups(["employee_get"])]
     private $city;
 
     /**
      * @var \DateTime Date of birth
-     * @ORM\Column(type="date")
-     * @Assert\NotNull
-     * @ApiProperty(
-     *     jsonldContext={"@type"="http://www.w3.org/2001/XMLSchema#date"}
-     * )
-     * @Groups({"employee_get", "employee_list"})
      */
+    #[ORM\Column (type:"date")]
+    #[Assert\NotNull]
+    #[Groups (["employee_get", "employee_list"])]
+    #[ApiProperty(jsonldContext: ['@type' => 'http://www.w3.org/2001/XMLSchema#date'])]
     private $birthDate;
 
     /**
      * @var \DateTime Time the employee usually arrives at work
-     * @ORM\Column(type="time", nullable=true)
-     * @ApiProperty(
-     *     jsonldContext={"@type"="http://www.w3.org/2001/XMLSchema#time"}
-     * )
-     * @Groups({"employee_get", "employee_list"})
      */
+    #[ORM\Column (type:"time", nullable:true)]
+    #[Groups (["employee_get", "employee_list"])]
+    #[ApiProperty(jsonldContext: ['@type' => 'http://www.w3.org/2001/XMLSchema#time'])]
     private $arrival;
 
     /**
      * @var Collection
-     * @ORM\OneToMany(targetEntity="App\Entity\Hours", mappedBy="employee")
      */
+    #[ORM\OneToMany(targetEntity:"App\Entity\Hours", mappedBy:"employee")]
     private $hours;
 
     /**
      * @var User associated with this employee
-     * @ORM\ManyToOne(targetEntity="App\Entity\User")
-     * @Groups({"employee_get"})
      */
+    #[ORM\ManyToOne(targetEntity:"App\Entity\User")]
+    #[Groups(["employee_get"])]
     private $user;
 
-    public function getId(): int
+    public function getId() : int
     {
         return $this->id;
     }
@@ -150,7 +143,7 @@ class Employee
     /**
      * @return string|null
      */
-    public function getFirstName(): ?string
+    public function getFirstName() : ?string
     {
         return $this->firstName;
     }
@@ -159,7 +152,7 @@ class Employee
      * @param string|null $firstName
      * @return Employee
      */
-    public function setFirstName(?string $firstName): Employee
+    public function setFirstName(?string $firstName) : Employee
     {
         $this->firstName = $firstName;
         return $this;
@@ -168,7 +161,7 @@ class Employee
     /**
      * @return string
      */
-    public function getLastName(): string
+    public function getLastName() : string
     {
         return $this->lastName;
     }
@@ -177,7 +170,7 @@ class Employee
      * @param string $lastName
      * @return Employee
      */
-    public function setLastName(string $lastName): Employee
+    public function setLastName(string $lastName) : Employee
     {
         $this->lastName = $lastName;
         return $this;
@@ -186,7 +179,7 @@ class Employee
     /**
      * @return string
      */
-    public function getJob(): string
+    public function getJob() : string
     {
         return $this->job;
     }
@@ -195,7 +188,7 @@ class Employee
      * @param string $job
      * @return Employee
      */
-    public function setJob(string $job): Employee
+    public function setJob(string $job) : Employee
     {
         $this->job = $job;
         return $this;
@@ -204,7 +197,7 @@ class Employee
     /**
      * @return string
      */
-    public function getAddress(): string
+    public function getAddress() : string
     {
         return $this->address;
     }
@@ -213,7 +206,7 @@ class Employee
      * @param string $address
      * @return Employee
      */
-    public function setAddress(string $address): Employee
+    public function setAddress(string $address) : Employee
     {
         $this->address = $address;
         return $this;
@@ -222,7 +215,7 @@ class Employee
     /**
      * @return string|null
      */
-    public function getZipcode(): ?string
+    public function getZipcode() : ?string
     {
         return $this->zipcode;
     }
@@ -231,7 +224,7 @@ class Employee
      * @param string $zipcode|null
      * @return Employee
      */
-    public function setZipcode(?string $zipcode): Employee
+    public function setZipcode(?string $zipcode) : Employee
     {
         $this->zipcode = $zipcode;
         return $this;
@@ -240,7 +233,7 @@ class Employee
     /**
      * @return string
      */
-    public function getCity(): string
+    public function getCity() : string
     {
         return $this->city;
     }
@@ -249,7 +242,7 @@ class Employee
      * @param string $city
      * @return Employee
      */
-    public function setCity(string $city): Employee
+    public function setCity(string $city) : Employee
     {
         $this->city = $city;
         return $this;
@@ -258,7 +251,7 @@ class Employee
     /**
      * @return \DateTime
      */
-    public function getBirthDate(): \DateTime
+    public function getBirthDate() : \DateTime
     {
         return $this->birthDate;
     }
@@ -267,7 +260,7 @@ class Employee
      * @param \DateTime $birthDate
      * @return Employee
      */
-    public function setBirthDate(\DateTime $birthDate): Employee
+    public function setBirthDate(\DateTime $birthDate) : Employee
     {
         $this->birthDate = $birthDate;
         return $this;
@@ -276,7 +269,7 @@ class Employee
     /**
      * @return \DateTime|null
      */
-    public function getArrival(): ?\DateTime
+    public function getArrival() : ?\DateTime
     {
         return $this->arrival;
     }
@@ -285,7 +278,7 @@ class Employee
      * @param \DateTime|null $arrival
      * @return Employee
      */
-    public function setArrival(\DateTime $arrival=null): Employee
+    public function setArrival(\DateTime $arrival = null) : Employee
     {
         $this->arrival = $arrival;
         return $this;
@@ -294,7 +287,7 @@ class Employee
     /**
      * @return Collection
      */
-    public function getHours(): Collection
+    public function getHours() : Collection
     {
         return $this->hours;
     }
@@ -308,21 +301,22 @@ class Employee
         $this->hours = $hours;
         return $this;
     }
-
     /**
      * Represent the entity to the user in a single string
+     *
      * @return string
-     * @ApiProperty(iri="http://schema.org/name")
-     * @Groups({"employee_get", "employee_list", "hours_get", "hours_list", "day_totals_per_employee"})
      */
-    function getLabel() {
-        return $this->getLastName(). ', '. $this->getFirstName();
+    #[Groups(["employee_get", "employee_list", "hours_get", "hours_list", "day_totals_per_employee"])]
+    #[ApiProperty(iris: ['http://schema.org/name'])]
+    function getLabel()
+    {
+        return $this->getLastName() . ', ' . $this->getFirstName();
     }
 
     /**
      * @return null|User
      */
-    public function getUser(): ?User
+    public function getUser() : ?User
     {
         return $this->user;
     }
@@ -331,10 +325,9 @@ class Employee
      * @param null|User $user
      * @return Employee
      */
-    public function setUser(User $user=null): Employee
+    public function setUser(User $user = null) : Employee
     {
         $this->user = $user;
         return $this;
     }
-
 }
