@@ -2,109 +2,107 @@
 
 namespace App\Entity;
 
-use ApiPlatform\Core\Annotation\ApiResource;
+use ApiPlatform\Metadata\Post;
+use ApiPlatform\Metadata\GetCollection;
+use ApiPlatform\Metadata\Delete;
+use ApiPlatform\Metadata\Put;
+use ApiPlatform\Metadata\Get;
+use ApiPlatform\Metadata\ApiResource;
+use ApiPlatform\Metadata\ApiProperty;
+use ApiPlatform\Doctrine\Orm\Filter\OrderFilter;
+use ApiPlatform\Doctrine\Orm\Filter\RangeFilter;
+use ApiPlatform\Doctrine\Orm\Filter\DateFilter;
+use ApiPlatform\Doctrine\Orm\Filter\SearchFilter;
+use ApiPlatform\Metadata\ApiFilter;
+use App\State\DayTotalsPerEmployeeCollectionProvider;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Validator\Constraints as Assert;
-use ApiPlatform\Core\Annotation\ApiProperty;
 use Symfony\Component\Serializer\Annotation\Groups;
-use ApiPlatform\Core\Annotation\ApiFilter;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\SearchFilter;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\DateFilter;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\RangeFilter;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\OrderFilter;
 use App\Validator\Constraints\CommonUserHoursStartConstraint;
 use App\Model\DayTotalsPerEmployee;
 
 /**
  * Registration of time worked by an Employee
- *
- * @ApiResource(attributes={
- *     "pagination_items_per_page"=10,
- *     "order"={"start": "DESC", "description": "ASC"},
- *     },
- *     itemOperations={
- *          "get"={
- *              "normalization_context"={"groups"={"hours_get"}},
- *              "security"="is_granted('ROLE_ADMIN') or object.getEmployee().getUser() == user"
- *          },
- *          "put"={"security_post_denormalize"="is_granted('ROLE_ADMIN') or
-               (object.getEmployee().getUser() == user and previous_object.getEmployee().getUser() == user)"},
- *          "delete"={"security"="is_granted('ROLE_ADMIN') or object.getEmployee().getUser() == user" }
- *     },
- *     collectionOperations={
- *         "get"={
- *              "normalization_context"={"groups"={"hours_list"}}
- *          },
- *         "post"={"security_post_denormalize"="is_granted('ROLE_ADMIN') or object.getEmployee().getUser() == user"},
- *         "get_day_report"={
- *             "method"="GET",
- *             "path"="/hours/dayreport",
- *             "output"=DayTotalsPerEmployee::class,
- *             "normalization_context"={
- *                  "groups"={"day_totals_per_employee"}},
- *             "pagination_enabled"=false
- *         }
- *     }
- * )
- * @ApiFilter(SearchFilter::class, properties={"description": "ipartial", "employee": "exact", "employee.job": "ipartial"})
- * @ApiFilter(DateFilter::class, properties={"start"})
- * @ApiFilter(RangeFilter::class, properties={"nHours"})
- * @ApiFilter(OrderFilter::class, properties={"start", "description", "nHours", "employee.firstName", "employee.lastName"})
- * @ORM\Entity
- * @ORM\Table(indexes={ @ORM\Index(columns={"start", "description"}) })
  */
+#[ORM\Entity]
+#[ORM\Table(indexes:[ new ORM\Index(columns: ["start", "description"]) ])]
+#[ApiResource(operations: [
+        new Get(normalizationContext: ['groups' => ['hours_get']],
+            security: 'is_granted(\'ROLE_ADMIN\') or object.getEmployee().getUser() == user',
+            requirements: ['id' => '\d+']),
+        new Put(securityPostDenormalize: 'is_granted(\'ROLE_ADMIN\') or
+                   (object.getEmployee().getUser() == user and previous_object.getEmployee().getUser() == user)'),
+        new Delete(security: 'is_granted(\'ROLE_ADMIN\') or object.getEmployee().getUser() == user'),
+        new GetCollection(normalizationContext: ['groups' => ['hours_list']]),
+        new Post(securityPostDenormalize: 'is_granted(\'ROLE_ADMIN\') or object.getEmployee().getUser() == user'),
+        new GetCollection(uriTemplate: '/hours/dayreport',
+            output: DayTotalsPerEmployee::class,
+            normalizationContext: ['groups' => ['day_totals_per_employee']],
+            paginationEnabled: false,
+            provider: DayTotalsPerEmployeeCollectionProvider::class,
+            openapiContext: [
+                'summary' => 'Totals per day per Employee',
+                'description' => 'Days start at time of start[after] filter',
+            ]
+        )
+    ],
+    paginationItemsPerPage: 10,
+    order: ['start' => 'DESC', 'description' => 'ASC'])
+]
+#[ApiFilter(filterClass: SearchFilter::class, properties: ['description' => 'ipartial', 'employee' => 'exact', 'employee.job' => 'ipartial'])]
+#[ApiFilter(filterClass: DateFilter::class, properties: ['start'])]
+#[ApiFilter(filterClass: RangeFilter::class, properties: ['nHours'])]
+#[ApiFilter(filterClass: OrderFilter::class, properties: ['start', 'description', 'nHours', 'employee.firstName', 'employee.lastName'])]
 class Hours
 {
     /**
      * @var int The entity Id
-     *
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
      */
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type:"integer")]
     private $id;
 
     /**
      * @var float number of hours
-     * @ORM\Column(type="float")
-     * @Assert\NotNull
-     * @Assert\GreaterThanOrEqual(0.1)
-     * @Groups({"hours_get", "hours_list"})
      */
+    #[ORM\Column(type:"float")]
+    #[Assert\NotNull]
+    #[Assert\GreaterThanOrEqual(0.1)]
+    #[Groups(["hours_get", "hours_list"])]
     private $nHours = 1.0;
 
     /**
      * @var \DateTime
-     *
-     * @ORM\Column(type="datetime")
-     * @Assert\NotNull
-     * @Groups({"hours_get", "hours_list"})
-     * @CommonUserHoursStartConstraint
      */
+    #[ORM\Column(type:"datetime")]
+    #[Assert\NotNull]
+    #[Groups(["hours_get", "hours_list"])]
+    #[CommonUserHoursStartConstraint]
     private $start;
 
     /**
      * @var bool
-     * @ORM\Column(type="boolean", nullable=true)
-     * @Groups({"hours_get"})
      */
+    #[ORM\Column(type:'boolean', nullable:true)]
+    #[Groups(["hours_get"])]
     private $onInvoice = true;
 
     /**
      * @var string
-     * @ORM\Column
-     * @Assert\NotBlank
-     * @Assert\Length(max=255)
-     * @Groups({"hours_get", "hours_list"})
      */
+    #[ORM\Column]
+    #[Assert\NotBlank]
+    #[Assert\Length(max:255)]
+    #[Groups(["hours_get", "hours_list"])]
     private $description;
 
     /**
      * @var Employee
-     * @ORM\ManyToOne(targetEntity="App\Entity\Employee", inversedBy="hours")
-     * @Assert\NotNull
-     * @Groups({"hours_get", "hours_list"})
      */
+    #[ORM\ManyToOne(targetEntity:"App\Entity\Employee", inversedBy:"hours")]
+    #[Assert\NotNull]
+    #[Groups(["hours_get", "hours_list"])]
     private $employee;
 
     public function __construct()
@@ -113,7 +111,7 @@ class Hours
         $this->setStart(new \DateTime());
     }
 
-    public function getId(): int
+    public function getId() : int
     {
         return $this->id;
     }
@@ -121,7 +119,7 @@ class Hours
     /**
      * @return float
      */
-    public function getNHours(): float
+    public function getNHours() : float
     {
         return $this->nHours;
     }
@@ -130,7 +128,7 @@ class Hours
      * @param float $nHours
      * @return Hours
      */
-    public function setNHours(float $nHours): Hours
+    public function setNHours(float $nHours) : Hours
     {
         $this->nHours = $nHours;
         return $this;
@@ -139,7 +137,7 @@ class Hours
     /**
      * @return \DateTime
      */
-    public function getStart(): \DateTime
+    public function getStart() : \DateTime
     {
         return $this->start;
     }
@@ -148,7 +146,7 @@ class Hours
      * @param \DateTime $start
      * @return Hours
      */
-    public function setStart(\DateTime $start): Hours
+    public function setStart(\DateTime $start) : Hours
     {
         $this->start = $start;
         return $this;
@@ -157,7 +155,7 @@ class Hours
     /**
      * @return bool
      */
-    public function isOnInvoice(): bool
+    public function isOnInvoice() : bool
     {
         return $this->onInvoice;
     }
@@ -166,7 +164,7 @@ class Hours
      * @param bool|null $onInvoice
      * @return Hours
      */
-    public function setOnInvoice(?bool $onInvoice): Hours
+    public function setOnInvoice(?bool $onInvoice) : Hours
     {
         $this->onInvoice = (bool) $onInvoice;
         return $this;
@@ -175,7 +173,7 @@ class Hours
     /**
      * @return string
      */
-    public function getDescription(): string
+    public function getDescription() : string
     {
         return $this->description;
     }
@@ -184,7 +182,7 @@ class Hours
      * @param string $description
      * @return Hours
      */
-    public function setDescription(string $description): Hours
+    public function setDescription(string $description) : Hours
     {
         $this->description = $description;
         return $this;
@@ -193,7 +191,7 @@ class Hours
     /**
      * @return Employee|null
      */
-    public function getEmployee(): ? Employee
+    public function getEmployee() : ?Employee
     {
         return $this->employee;
     }
@@ -208,22 +206,24 @@ class Hours
         return $this;
     }
 
-
-    /** Represent the entity to the user in a single string
-     * @ApiProperty(iri="http://schema.org/name")
-     * @Groups({"hours_get", "hours_list"})
+    /**
+     * Represent the entity to the user in a single string
+     *
      * @return string
      */
-    public function getLabel() {
-        return $this->getStart()->format('Y-m-d H:i:s')
-            . ' '. $this->getDescription();
+    #[Groups (["hours_get", "hours_list"])]
+    #[ApiProperty(iris: ['http://schema.org/name'])]
+    public function getLabel()
+    {
+        return $this->getStart()->format('Y-m-d H:i:s') . ' ' . $this->getDescription();
     }
 
     /**
      * @return string
-     * @Groups({"hours_get", "hours_list"})
      */
-    public function getDay() {
+    #[Groups (["hours_get", "hours_list"])]
+    public function getDay()
+    {
         return $this->getStart()->format('D');
     }
 }
