@@ -2,12 +2,12 @@
 
 namespace App\Filter;
 
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\AbstractContextAwareFilter;
-use ApiPlatform\Core\Bridge\Doctrine\Orm\Util\QueryNameGeneratorInterface;
+use ApiPlatform\Doctrine\Orm\Filter\AbstractFilter;
+use ApiPlatform\Doctrine\Orm\Util\QueryNameGeneratorInterface;
+use ApiPlatform\Metadata\Operation;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 use ApiPlatform\Core\Exception\InvalidArgumentException;
 
@@ -19,7 +19,7 @@ use ApiPlatform\Core\Exception\InvalidArgumentException;
  * All specified properties type must be string.
  * @package App\Filter
  */
-class SimpleSearchFilter extends AbstractContextAwareFilter
+class SimpleSearchFilter extends AbstractFilter
 {
     private $searchParameterName;
 
@@ -28,15 +28,15 @@ class SimpleSearchFilter extends AbstractContextAwareFilter
      * {@inheritdoc}
      * @param string $searchParameterName The parameter whose value this filter searches for
      */
-    public function __construct(ManagerRegistry $managerRegistry, ?RequestStack $requestStack = null, LoggerInterface $logger = null, array $properties = null, NameConverterInterface $nameConverter = null, string $searchParameterName = 'simplesearch')
+    public function __construct(ManagerRegistry $managerRegistry, LoggerInterface $logger = null, array $properties = null, NameConverterInterface $nameConverter = null, string $searchParameterName = 'simplesearch')
     {
-        parent::__construct($managerRegistry, $requestStack, $logger, $properties, $nameConverter);
+        parent::__construct($managerRegistry, $logger, $properties, $nameConverter);
 
         $this->searchParameterName = $searchParameterName;
     }
 
     /** {@inheritdoc} */
-    protected function filterProperty(string $property, $value, QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, string $operationName = null, array $context = [])
+    protected function filterProperty(string $property, $value, QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, Operation $operation = null, array $context = []): void
     {
         if (null === $value || $property !== $this->searchParameterName) {
             return;
